# openapi.yaml
openapi: 3.1.0
info:
  title: Ticket API
  version: 0.1.891
  description: |
    RESTful API for the Ticket project and issue management system.
    All endpoints are under `/api/`. Authentication uses Bearer tokens
    or session cookies (`__Host-session` on secure requests; legacy
    `ticket_token` remains accepted).
    Agent endpoints use HTTP Basic Auth (agent UUID + password).
  contact:
    name: Simon Gauld
    url: https://github.com/simonski/ticket
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token from /api/login
    CookieAuth:
      type: apiKey
      in: cookie
      name: __Host-session
    BasicAuth:
      type: http
      scheme: basic
      description: Agent UUID as username, agent password as password
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
      example:
        error: "ticket not found"
...
    Project:
      type: object
      properties:
        project_id:
          type: integer
          format: int64
        prefix:
          type: string
          description: 1-5 uppercase ASCII letters
          pattern: '^[A-Z]{1,5}$'
        title:
          type: string
        description:
          type: string
        acceptance_criteria:
          type: string
        git_repository:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [open, closed]
        visibility:
          type: string
          enum: [public, private]
        workflow_id:
          type: integer
          format: int64
          nullable: true
        ticket_sequence:
          type: integer
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Label:
      type: object
      properties:
        label_id:
          type: integer
          format: int64
        project_id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string
        created_at:
          type: string
          format: date-time
...
    Document:
      type: object
      properties:
        document_id:
          type: integer
          format: int64
        project_id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        notes:
          type: string
        content:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DocumentFile:
      type: object
      properties:
        file_id:
          type: integer
          format: int64
        document_id:
          type: integer
          format: int64
        file_name:
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer
          format: int64
        content:
          type: string
          format: byte
        created_at:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  examples:
...
    ProjectExample:
      summary: A typical project
      value:
        project_id: 1
        prefix: "CUS"
        title: "Customer Portal"
        description: "Customer-facing web portal project"
        status: "open"
        visibility: "private"
        created_at: "2025-01-15T10:00:00Z"
        updated_at: "2025-06-20T14:30:00Z"
...
    ProjectListExample:
      summary: Project list response
      value:
        - project_id: 1
          prefix: "CUS"
          title: "Customer Portal"
          description: "Customer-facing web portal project"
          status: "open"
          visibility: "private"
        - project_id: 2
          prefix: "OPS"
          title: "Operations Console"
          description: "Internal operational tooling"
          status: "open"
          visibility: "private"
    CreateProjectRequestExample:
      summary: Create project request
      value:
        prefix: "CUS"
        title: "Customer Portal"
        description: "Customer-facing web portal project"
        visibility: "private"
        workflow_id: 1
...
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List projects visible to the authenticated user
      operationId: listProjects
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
              examples:
                default:
                  $ref: '#/components/examples/ProjectListExample'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Projects]
      summary: Create a project
      operationId: createProject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prefix, title]
              properties:
                prefix:
                  type: string
                  description: 1-5 uppercase ASCII letters
                  pattern: '^[A-Z]{1,5}$'
                title:
                  type: string
                description:
                  type: string
                acceptance_criteria:
                  type: string
                git_repository:
                  type: string
                notes:
                  type: string
                visibility:
                  type: string
                  enum: [public, private]
                workflow_id:
                  type: integer
                  format: int64
            examples:
              default:
                $ref: '#/components/examples/CreateProjectRequestExample'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              examples:
                created:
                  $ref: '#/components/examples/ProjectExample'

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/projects/{project_id}:
    get:
      tags: [Projects]
      summary: Get project details
      operationId: getProject
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              examples:
                default:
                  $ref: '#/components/examples/ProjectExample'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Projects]
      summary: Update a project
      operationId: updateProject
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                acceptance_criteria:
                  type: string
                git_repository:
                  type: string
                notes:
                  type: string
                status:
                  type: string
                  enum: [open, closed]
                visibility:
                  type: string
                  enum: [public, private]
                workflow_id:
                  type: integer
                  format: int64
            examples:
              default:
                value:
                  title: "Customer Portal"
                  description: "Customer-facing web portal project"
                  visibility: "private"
                  workflow_id: 1
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              examples:
                updated:
                  $ref: '#/components/examples/ProjectExample'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/projects/{project_id}/documents:
    get:
      tags: [Projects, Documents]
      summary: List project documents
      operationId: listProjectDocuments
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Projects, Documents]
      summary: Create a project document
      operationId: createProjectDocument
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                notes:
                  type: string
                content:
                  type: string
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/documents/{document_id}:
    get:
      tags: [Documents]
      summary: Get a document
      operationId: getDocument
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Documents]
      summary: Update a document
      operationId: updateDocument
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                notes:
                  type: string
                content:
                  type: string
      responses:
        '200':
          description: Document updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Documents]
      summary: Delete a document
      operationId: deleteDocument
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Document deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/documents/{document_id}/labels:
    get:
      tags: [Documents, Labels]
      summary: List labels on a document
      operationId: listDocumentLabels
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Document labels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Documents, Labels]
      summary: Add a label to a document
      operationId: addDocumentLabel
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label_id]
              properties:
                label_id:
                  type: integer
                  format: int64
      responses:
        '204':
          description: Label added
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Documents, Labels]
      summary: Remove a label from a document
      operationId: removeDocumentLabel
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: label_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Label removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/documents/{document_id}/files:
    get:
      tags: [Documents]
      summary: List files on a document
      operationId: listDocumentFiles
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Document file metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentFile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Documents]
      summary: Add a file to a document
      operationId: addDocumentFile
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_name, content]
              properties:
                file_name:
                  type: string
                content_type:
                  type: string
                content:
                  type: string
                  format: byte
      responses:
        '201':
          description: Document file created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentFile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/documents/{document_id}/files/{file_id}:
    get:
      tags: [Documents]
      summary: Download a document file
      operationId: getDocumentFile
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: file_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: File bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Documents]
      summary: Delete a document file
      operationId: deleteDocumentFile
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: file_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Document file deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  # ── Stories ─────────────────────────────────────────────────────────
...